if statement is added to avoid errors, when testing. Not sure 100% is this is a solution or workaround

+++ Lib/tempfile.py
@@ -444,7 +444,8 @@ class _TemporaryFileCloser:
                     self.file.close()
                 finally:
                     if self.delete and self.delete_on_close:
-                        unlink(self.name)
+                        if _os.path.exists(self.name):
+                            unlink(self.name)
 
         # Need to ensure the file is deleted on __del__
         def __del__(self):
